[connectors] - refactor: streamline Slack channel creation webhook handler
- Replace manual Slack channel joining logic with `autoJoinChannel` function for clarity and maintainability
- Remove unused imports and nested conditionals simplifying the `channel_created` event case
- Clean up code formatting by ensuring consistent trailing commas and removal of superfluous whitespace
